% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{changeConfiguration}
\alias{changeConfiguration}
\title{changeConfigutation allows changes at E2Predictor configuration}
\usage{
changeConfiguration(netMHCpath = E2Predictor.Config$netMHCpath,
  working.path = E2Predictor.Config$working.path,
  hla_alleles = E2Predictor.Config$hla_alleles)
}
\arguments{
\item{netMHCpath}{path to netMHC 4.0 (only necessary if new allele predictions are needed)}

\item{working.path}{by default the current working path}

\item{hla_alleles}{list of cell.line - alleles}
}
\description{
changeConfigutation allows changes at E2Predictor configuration
}

