% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.learning.R
\name{contingencyTable}
\alias{contingencyTable}
\title{contingencyTable}
\usage{
contingencyTable(predicted, test_var, threshold = 0.6)
}
\arguments{
\item{predicted}{vector with predicted values (if it is a score, a threshold should be given)}

\item{test_var}{vector with classification}
}
\value{
data frame with the four values of the contingency table
}
\description{
calculates the contingency table, given test and predicted values, and a threshold for the predicted values.
}

