% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countCharOccurrences}
\alias{countCharOccurrences}
\title{countCharOccurrences counts the presence of a given character into a string}
\usage{
countCharOccurrences(s, char)
}
\arguments{
\item{s}{string where chars should be counted}

\item{char}{character you want to count}
}
\description{
countCharOccurrences counts the presence of a given character into a string
}

