% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.learning.R
\name{evaluate_ML}
\alias{evaluate_ML}
\title{evaluate_ML}
\usage{
evaluate_ML(predictor, testset, class_category, save.to = NULL)
}
\arguments{
\item{predictor}{ML predict output object}

\item{testset}{data set for testing}

\item{class_category}{classification category}

\item{save.to}{path to save results. When NULL, results will be displayed online}
}
\description{
evaluate a Machine Learning prediction
}

