% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.predictors.R
\name{exec.netMHCpan}
\alias{exec.netMHCpan}
\title{exec.netMHCpan executes the command netMHCpan for a set of peptides of SAME length and ONE allele (a single prediction)}
\usage{
exec.netMHCpan(allele, peptides,
  outputPath = file.path(E2Predictor.Config$working.path, "temp"))
}
\arguments{
\item{allele}{allele name (defined as of E2Predictor.Config$hla_alleles)}

\item{outputPath}{}

\item{peptide}{peptide sequence string}
}
\description{
exec.netMHCpan executes the command netMHCpan for a set of peptides of SAME length and ONE allele (a single prediction)
}

