% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.learning.R
\name{get_training_sample}
\alias{get_training_sample}
\title{get.training.sample}
\usage{
get_training_sample(df, class_category, percentage_training)
}
\arguments{
\item{df}{data frame containing a whole data set.}

\item{class_category}{boolean column name used for classification.}

\item{percentage_training}{percentage of data used for training. It takes the portion percentage_training from the smallest class and a equivalent number of elements from the biggest class.}
}
\value{
list with two data frames: trainset and testset
}
\description{
separates a data frame into two data frames for training and testing in ML classification.
}

