% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{initConfiguration}
\alias{initConfiguration}
\title{initConfiguration initializes E2Predictor configuration}
\usage{

  initConfiguration(netMHCpath = "/Users/napedro/external_sources/netMHC-4.0/netMHC",
  working.path = getwd(), hla_alleles = list(JY = c("A*02:01", "B*07:02",
  "C*07:02"), LCLC.103H = c("A*03:01", "B*07:02", "C*07:02"), COLO.699.N =
  c("A*02:01", "B*07:02", "C*07:02", "C*06:02"), MEL.526 = c("A*02:01",
  "B*15:01", "C*07:02"), HEK293 = c("A*03:01", "B*07:02", "C*07:02"), MEL.HO =
  c("A*03:01", "B*07:02", "B*14:02", "C*07:02", "C*08:05"), NIH.OVCAR.3 =
  c("A*02:01", "A*29:02", "B*07:02", "B*58:01", "C*07:02", "C*07:19"), SK.MEL.37
  = c("A*02:01", "B*07:02", "B*56:01", "C*07:02", "C*03:04"), K562A2 =
  c("A*02:01", "C*05:01", "C*03:04")))
}
\arguments{
\item{netMHCpath}{path to netMHC 4.0 (only necessary if new allele predictions are needed)}

\item{working.path}{by default the current working path}

\item{hla_alleles}{list of cell.line - alleles}
}
\description{
initConfiguration initializes E2Predictor configuration
}

