% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.files.R
\name{load_data_modeling}
\alias{load_data_modeling}
\title{load_data_modeling}
\usage{
load_data_modeling(cell.line, is_perturbation, condition = "",
  addgrepcond = "", requireAllele = F)
}
\arguments{
\item{cell.line}{cell line to be loaded}

\item{is_perturbation}{data from perturbations?}

\item{condition}{if data from perturbations, condition to be loaded}

\item{addgrepcond}{an extra parameter to locate the right condition. Example: to distinguish conditions like "Control I" and "Control II"}

\item{requireAllele}{proteins with ligands require at least one ligand associated to an allele.}
}
\value{
a data.frame containing data for modeling
}
\description{
loads a data modeling table
}

