% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.omics.R
\name{merge.omics}
\alias{merge.omics}
\title{merge.omics merges proteomics, genomics, and ligandomics data of an experiment}
\usage{
\method{merge}{omics}(selected.cell.line, dbfile, is_perturbation = F,
  condition = NULL, additionalGrepCondition = "", minReplicates = 1,
  allele.predictions = NULL, allele.predictor = "ic50",
  nM.threshold = 1000, saveAllResults = F, GOcategories = NULL)
}
\arguments{
\item{selected.cell.line}{cell line to integrate data}

\item{dbfile}{Uniprot database filename (NOT in fasta format! Use the xls format)}

\item{is_perturbation}{indicates if the experiment is a perturbation experiment}

\item{condition}{in case of a perturbation, this indicates which condition you want to use}

\item{minReplicates}{for ligandomics data, the minimum number of biological replicates required to validate a ligand}

\item{allele.predictions}{data frame containing allele predictions for the ligands. If NULL, they will be estimated}

\item{allele.predictor}{c("rank", "ic50") predictor to be used.}

\item{nM.threshold}{minimum nanomolar threshold for validating the predictions}

\item{saveAllResults}{save allele predictions and merged omics table at the working directory}

\item{GOcategories}{a string vector containing gene ontology categories you want to include as separate variables}
}
\description{
merge.omics merges proteomics, genomics, and ligandomics data of an experiment
}

