% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.predictors.R
\name{predictNeoEpitopes}
\alias{predictNeoEpitopes}
\title{predictNeoEpitopes predicts the most likely alleles for ligands}
\usage{
predictNeoEpitopes(ligands, predictor.value = c("rank", "ic50"),
  seq.min.length = 8, seq.max.length = 14)
}
\arguments{
\item{ligands}{a data frame, which should contain at least two variables: Sequence, cell.line}

\item{predictor.value}{indicates which value is used to estimate the best prediction between different alleles}
}
\value{
netMHCpredictions data.frame containing allele predictions
}
\description{
predictNeoEpitopes predicts the most likely alleles for ligands
}

