% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.learning.R
\name{prelim.ML}
\alias{prelim.ML}
\title{prelim.ML}
\usage{
prelim.ML(df, ncp = 5, dim.reduction.factor = 0.7, saveInFolder = NULL,
  suffix = "")
}
\arguments{
\item{df}{data frame for analysis}

\item{ncp}{number of dimensions kept in the results (by default 5) of PCA analysis}

\item{dim.reduction.factor}{Correlation threshold over which correlated variables will be dropped}

\item{saveInFolder}{when not NULL, folder path where result files will be written}

\item{suffix}{when saveInFolder is not NULL, it adds a suffix to file names}
}
\value{
list with max four objects: PCA and correlation of variables, and if variable reduction is possible, PCA and correlation of reduced variables.
}
\description{
makes some preliminary steps useful for Machine Learning (A PCA and a correlation analysis)
}

