% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.learning.R
\name{prepare_data_modeling}
\alias{prepare_data_modeling}
\title{prepare_data_modeling}
\usage{
prepare_data_modeling(df, useVariables = NULL)
}
\arguments{
\item{df}{data frame containing data to be modeled}

\item{useVariables}{a vector containing variable names to be used. When NULL, all variables are used.}
}
\description{
fix some variable names (from gene ontology), filter variables, and normalize data
}

