% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.files.R
\name{proteomics.select.condition}
\alias{proteomics.select.condition}
\title{proteomics.select.condition selects the quantitative data from a proteomics (ISOQuant) data frame}
\usage{
proteomics.select.condition(proteindf, condition, useAverage = T)
}
\arguments{
\item{proteindf}{data frame containing Proteomics data (ISOQuant format)}

\item{condition}{condition or name you want to use to filter data}

\item{useAverage}{if TRUE, averaged values of columns found under 'condition' will be returned}
}
\description{
proteomics.select.condition selects the quantitative data from a proteomics (ISOQuant) data frame
}

