% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_batch_in_folder}
\alias{read_batch_in_folder}
\title{read_batch_in_folder reads several files from a folder, and concatenates them}
\usage{
read_batch_in_folder(folder, filepattern = ".", keepFolderName = F,
  UseGenericHeaders = T)
}
\arguments{
\item{folder}{folder where files will be read}

\item{filepattern}{regular expression to filter files in the folder by file name}

\item{keepFolderName}{keeps the folder name in a column instead of the file name}

\item{UseGenericHeaders}{substitute headers by generic headers (useful when concatenating several files and headers are different)}
}
\description{
read_batch_in_folder reads several files from a folder, and concatenates them
}

