% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_file_add_filename}
\alias{read_file_add_filename}
\title{read_file_add_filename reads a file and attaches a column with the file name or the folder name}
\usage{
read_file_add_filename(thefile, keepFolderName = F, UseGenericHeaders = F,
  ...)
}
\arguments{
\item{thefile}{the file to be read}

\item{keepFolderName}{uses the folder name instead of the file name}

\item{UseGenericHeaders}{substitute headers by generic headers (useful when concatenating several files and headers are different)}
}
\description{
read_file_add_filename reads a file and attaches a column with the file name or the folder name
}

