% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.files.R
\name{read_file_crawler}
\alias{read_file_crawler}
\title{read_file_crawler reads a file produced by crawler script and
prepares the table to be used for modeling}
\usage{
read_file_crawler(thefile, ...)
}
\arguments{
\item{thefile}{crawler results file}
}
\description{
read_file_crawler reads a file produced by crawler script and
prepares the table to be used for modeling
}

