% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.files.R
\name{read_file_protein}
\alias{read_file_protein}
\title{read_file_protein reads a file with protein readings}
\usage{
read_file_protein(thefile, ...)
}
\arguments{
\item{thefile}{file with protein readings}
}
\description{
read_file_protein reads a file with protein readings
}

