% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.files.R
\name{read_file_RPKM}
\alias{read_file_RPKM}
\title{read_file_RPKM reads a file with RPKM readings of a SINGLE measurement (2 columns!)}
\usage{
read_file_RPKM(thefiles, useAverage = FALSE, minValueRPKM = 0, ...)
}
\arguments{
\item{thefiles}{vector of file name(s) containing RPKM readings}

\item{useAverage}{if TRUE, averaged values of all file names provided are returned}
}
\description{
read_file_RPKM reads a file with RPKM readings of a SINGLE measurement (2 columns!)
}

