% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.files.R
\name{read_ligands_PEAKS}
\alias{read_ligands_PEAKS}
\title{read_ligands_PEAKS reads and combines PEAKS identifications of PEAKS results folders.}
\usage{
read_ligands_PEAKS(PEAKSresultsFolders, minReplicates = 1,
  ligands_min_length = 8, ligands_max_length = 14, uniprotDB = NULL,
  alleles.prediction = NULL)
}
\arguments{
\item{PEAKSresultsFolders}{folders containing PEAKS results}

\item{minReplicates}{minimum number of replicates (in different PEAKS results folders) desired for each identified peptide}

\item{ligands_min_length}{minimum length of ligand sequences}

\item{ligands_max_length}{maximum length of ligand sequences}
}
\description{
read_ligands_PEAKS reads and combines PEAKS identifications of PEAKS results folders.
}

