% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_the_file}
\alias{read_the_file}
\title{read_the_file is a wrapper of read_excel and read_delim (from readr and readxl packages) that
checks which is the proper delimiter for the file before reading it.}
\usage{
read_the_file(thefile, ...)
}
\arguments{
\item{thefile}{file to be read}
}
\description{
read_the_file is a wrapper of read_excel and read_delim (from readr and readxl packages) that
checks which is the proper delimiter for the file before reading it.
}

