% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.learning.R
\name{retrieve_data_modeling}
\alias{retrieve_data_modeling}
\title{retrieve.data.modeling}
\usage{
retrieve_data_modeling(cell.line, is_perturbation, ML_method)
}
\arguments{
\item{cell.line}{cell line from which data should be retrieved}
}
\description{
retrieves data for modeling from a defined data structure
}

